requirements.txt
pymongo==4.2.0
pandas
numpy
scikit-learn
pymongo
python-dotenv
PyYAML
dill
scipy
imblearn==0.0

-e .